requirements.txt
nptyping==2.5.0
numpy==1.24.3
Pillow==9.5.0
pynrrd==1.0.0
tqdm==4.65.0
typing_extensions==4.5.0
